requirements.txt
# MonoVista - Python dependencies
# Python >= 3.9

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
gunicorn>=21.2.0
PyMySQL>=1.1.1
PyJWT>=2.8.0
oss2>=2.18.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
bleach>=6.1.0
user-agents>=2.2.0
